# doc/apidoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Khazana POS API",
    "description": "## Khazana POS Backend API\n\nAvailable modules:\n- **Auth** – Login, profile, change password\n- **Users** – User management and role assignment\n- **Sales** – Cash, credit, VAT cash, VAT credit sales\n- **Inventory** – Items, stock receive / issue / adjust\n- **Customers** – Customer management, ledger, payments\n- **Roles** – Role management\n- **Menus** – Menu items\n- **Permissions** – Role-based permissions\n- **Orders** – Sales orders (regular and VAT)\n- **Pricing** – Sale prices and cost prices\n- **Finance** – Money receive and cash purchases\n- **Packets** – Packet stock management\n- **NC Adjustment** – Non-conformance adjustments\n- **Assortment** – Assortment records\n- **Reports** – Sales, stock, item-wise, customer statements\n- **Admin** – Branches, settings, banks, audit log\n- **Dashboard** – Summary statistics\n- **Categories** – Item category management",
    "version": "1.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPreviousPage": { "type": "boolean", "example": false }
        }
      },
...
    }
  },
  "paths": {
...
    "/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get all customers",
        "operationId": "CustomersController_findAll",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Customer retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Create a new customer",
        "operationId": "CustomersController_create",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Customer created successfully" },
          "409": { "description": "Customer code already exists" }
        }
      }
    },
    "/customers/{code}": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get customer by code",
        "operationId": "CustomersController_findOne",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "responses": {
          "200": { "description": "Customer found" },
          "404": { "description": "Customer not found" }
        }
      },
      "patch": {
        "tags": ["Customers"],
        "summary": "Update customer by code",
        "operationId": "CustomersController_update",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "200": { "description": "Customer updated successfully" }
        }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Delete customer by code",
        "operationId": "CustomersController_remove",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "responses": {
          "200": { "description": "Customer deleted successfully" },
          "404": { "description": "Customer not found" }
        }
      }
    },
    "/customers/payments": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get all customer payments (register)",
        "operationId": "CustomersController_findAllPayments",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of payments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Payment retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Record a payment (client code in body)",
        "operationId": "CustomersController_createPayment",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Payment recorded successfully" }
        }
      }
    },
    "/customers/{code}/ledger": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get customer ledger (all transactions)",
        "operationId": "CustomersController_getLedger",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "responses": {
          "200": { "description": "Customer ledger entries" }
        }
      }
    },
    "/customers/{code}/balance": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get customer outstanding balance",
        "operationId": "CustomersController_getBalance",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "responses": {
          "200": { "description": "Customer balance summary" }
        }
      }
    },
    "/customers/{code}/payments": {
      "post": {
        "tags": ["Customers"],
        "summary": "Record a payment from customer",
        "operationId": "CustomersController_addPayment",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Payment recorded successfully" }
        }
      },
      "get": {
        "tags": ["Customers"],
        "summary": "Get all payments from customer",
        "operationId": "CustomersController_findPayments",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Customer code" }
        ],
        "responses": {
          "200": { "description": "List of payments" }
        }
      }
    },
...
    "/finance/money-receive": {
      "get": {
        "tags": ["Finance"],
        "summary": "Get all money receive records",
        "operationId": "FinanceController_findAllMoneyReceive",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of money receive records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Money Receive retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Finance"],
        "summary": "Record a money receipt from customer",
        "operationId": "FinanceController_createMoneyReceive",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Money receive recorded successfully" }
        }
      }
    },
    "/finance/cash-purchase": {
      "get": {
        "tags": ["Finance"],
        "summary": "Get all cash purchase records",
        "operationId": "FinanceController_findAllCashPurchases",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of cash purchases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Cash Purchase retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Finance"],
        "summary": "Record a cash purchase",
        "operationId": "FinanceController_createCashPurchase",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object" } }
          }
        },
        "responses": {
          "201": { "description": "Cash purchase recorded successfully" }
        }
      }
    },
...
  },
...
}